--- tools/__tasks__/compile/data/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-    description: 'Clean download and build data assets',
-    task: [require('./clean.mjs'), require('./download.mjs'), require('./amp.mjs')],
-};
