+++ tools/__tasks__/compile/data/index.mjs
@@ -1,5 +1,5 @@
 export default {
-	description: 'Clean download and build data assets (dev)',
+	description: 'Clean download and build data assets',
 	task: [
 		import('./clean.mjs'),
 		import('./download.mjs'),
