--- tools/__tasks__/compile/hash/clean.js
@@ -1,9 +1,0 @@
-const path = require('path');
-const rimraf = require('rimraf');
-
-const { hash } = require('../.././config.mjs').paths;
-
-module.exports = {
-    description: 'Clear asset hash artefacts',
-    task: () => rimraf.sync(path.resolve(hash, 'assets')),
-};
