+++ tools/__tasks__/compile/index.dev.js
@@ -1,10 +1,10 @@
-module.exports = {
-    description: 'Compile assets for development',
-    task: [
-        require('./conf/clean.mjs'),
-        require('./css/index.dev.mjs'),
-        require('./data/index.dev'),
-        require('./javascript/index.dev'),
-        require('./conf/index.mjs'),
-    ],
+export default {
+	description: 'Compile assets for development',
+	task: [
+		import('./conf/clean.mjs'),
+		import('./css/index.dev.mjs'),
+		import('./data/index.dev'),
+		import('./javascript/index.dev'),
+		import('./conf/index.mjs'),
+	],
 };
