--- tools/__tasks__/compile/javascript/index.atoms.js
@@ -1,8 +1,0 @@
-module.exports = {
-    description: 'Compile JS',
-    task: [
-        require('./clean.mjs'),
-        require('../inline-svgs/index.mjs'),
-        require('./webpack-atoms'),
-    ],
-};
