+++ tools/__tasks__/compile/javascript/index.atoms.mjs
@@ -1,0 +1,8 @@
+export default {
+	description: 'Compile JS',
+	task: [
+		import('./clean.mjs'),
+		import('../inline-svgs/index.mjs'),
+		import('./webpack-atoms.mjs'),
+	],
+};
