--- tools/__tasks__/compile/javascript/index.dev.js
@@ -1,12 +1,0 @@
-module.exports = {
-    description: 'Prepare JS for development',
-    task: [
-        require('../inline-svgs/index.mjs'),
-        require('./clean.mjs'),
-        require('./copy.mjs'),
-        require('../../commercial/compile'),
-        require('./webpack.dev'),
-        require('./webpack-dcr.dev'),
-        require('./bundle-polyfills.mjs'),
-    ],
-};
