--- tools/__tasks__/compile/javascript/index.js
@@ -1,11 +1,0 @@
-module.exports = {
-    description: 'Compile JS',
-    task: [
-        require('./clean.mjs'),
-        require('../inline-svgs/index.mjs'),
-        require('./copy.mjs'),
-        require('./webpack'),
-        require('./webpack-atoms'),
-        require('./bundle-polyfills.mjs'),
-    ],
-};
