+++ tools/__tasks__/compile/javascript/index.mjs
@@ -1,0 +1,11 @@
+export default {
+	description: 'Compile JS',
+	task: [
+		import('./clean.mjs'),
+		import('../inline-svgs/index.mjs'),
+		import('./copy.mjs'),
+		import('./webpack.mjs'),
+		import('./webpack-atoms.mjs'),
+		import('./bundle-polyfills.mjs'),
+	],
+};
