--- tools/__tasks__/compile/javascript/webpack-atoms.js
@@ -1,33 +1,0 @@
-
-require('any-observable/register/rxjs-all');
-
-const Observable = require('any-observable');
-
-const webpack = require('webpack');
-const chalk = require('chalk');
-
-const config = require('../../../../webpack.config.atoms.js');
-
-module.exports = {
-    description: 'Create Webpack bundles for atoms',
-    task: () =>
-        new Observable(observer => {
-            config.plugins = [
-                require('../../../webpack-progress-reporter')(observer),
-                ...config.plugins,
-            ];
-
-            const bundler = webpack(config);
-
-            bundler.run((err, stats) => {
-                if (err) {
-                    throw new Error(chalk.red(err));
-                }
-                const info = stats.toJson();
-                if (stats.hasErrors()) {
-                    throw new Error(chalk.red(info.errors));
-                }
-                observer.complete();
-            });
-        }),
-};
